django_api/pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["Grzegorz Dlugokecki <grzesiek.dlugokecki@gmail.com>"]
package-mode=false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
django-ninja = "^1.3.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
email-validator = "^2.2.0"
django-ninja-jwt = "^5.3.5"
factory-boy = "^3.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
